Refactor pkg/binding and separate add and remove code
