+++ pkg/binding/add_binding.go
@@ -1,0 +1,116 @@
+package binding
+
+import (
+	"fmt"
+
+	"github.com/devfile/library/pkg/devfile/parser"
+	sboApi "github.com/redhat-developer/service-binding-operator/apis/binding/v1alpha1"
+	"gopkg.in/yaml.v2"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+
+	backendpkg "github.com/redhat-developer/odo/pkg/binding/backend"
+	"github.com/redhat-developer/odo/pkg/kclient"
+	"github.com/redhat-developer/odo/pkg/libdevfile"
+)
+
+// ValidateAddBinding calls Validate method of the adequate backend
+func (o *BindingClient) ValidateAddBinding(flags map[string]string) error {
+	var backend backendpkg.AddBindingBackend
+	if len(flags) == 0 {
+		backend = o.interactiveBackend
+	} else {
+		backend = o.flagsBackend
+	}
+	return backend.Validate(flags)
+}
+
+func (o *BindingClient) SelectServiceInstance(flags map[string]string, serviceMap map[string]unstructured.Unstructured) (string, error) {
+	var backend backendpkg.AddBindingBackend
+	if len(flags) == 0 {
+		backend = o.interactiveBackend
+	} else {
+		backend = o.flagsBackend
+	}
+	return backend.SelectServiceInstance(flags[backendpkg.FLAG_SERVICE], serviceMap)
+}
+
+func (o *BindingClient) AskBindingName(serviceName, componentName string, flags map[string]string) (string, error) {
+	var backend backendpkg.AddBindingBackend
+	if len(flags) == 0 {
+		backend = o.interactiveBackend
+	} else {
+		backend = o.flagsBackend
+	}
+	defaultBindingName := fmt.Sprintf("%v-%v", componentName, serviceName)
+	return backend.AskBindingName(defaultBindingName, flags)
+}
+
+func (o *BindingClient) AskBindAsFiles(flags map[string]string) (bool, error) {
+	var backend backendpkg.AddBindingBackend
+	if len(flags) == 0 {
+		backend = o.interactiveBackend
+	} else {
+		backend = o.flagsBackend
+	}
+	return backend.AskBindAsFiles(flags)
+}
+
+func (o *BindingClient) AddBinding(bindingName string, bindAsFiles bool, unstructuredService unstructured.Unstructured, obj parser.DevfileObj, componentContext string) (parser.DevfileObj, error) {
+	service, err := o.kubernetesClient.NewServiceBindingServiceObject(unstructuredService, bindingName)
+	if err != nil {
+		return obj, err
+	}
+
+	deploymentName := fmt.Sprintf("%s-app", obj.GetMetadataName())
+	deploymentGVR, err := o.kubernetesClient.GetDeploymentAPIVersion()
+	if err != nil {
+		return obj, err
+	}
+
+	serviceBinding := kclient.NewServiceBindingObject(bindingName, bindAsFiles, deploymentName, deploymentGVR, []sboApi.Mapping{}, []sboApi.Service{service})
+
+	// Note: we cannot directly marshal the serviceBinding object to yaml because it doesn't do that in the correct k8s manifest format
+	serviceBindingUnstructured, err := kclient.ConvertK8sResourceToUnstructured(serviceBinding)
+	if err != nil {
+		return obj, err
+	}
+	yamlDesc, err := yaml.Marshal(serviceBindingUnstructured.UnstructuredContent())
+	if err != nil {
+		return obj, err
+	}
+
+	return libdevfile.AddKubernetesComponentToDevfile(string(yamlDesc), serviceBinding.Name, obj)
+}
+
+func (o *BindingClient) GetServiceInstances() (map[string]unstructured.Unstructured, error) {
+	// Get the BindableKinds/bindable-kinds object
+	bindableKind, err := o.kubernetesClient.GetBindableKinds()
+	if err != nil {
+		return nil, err
+	}
+
+	// get a list of restMappings of all the GVKs present in bindableKind's Status
+	bindableKindRestMappings, err := o.kubernetesClient.GetBindableKindStatusRestMapping(bindableKind.Status)
+	if err != nil {
+		return nil, err
+	}
+
+	var bindableObjectMap = map[string]unstructured.Unstructured{}
+	for _, restMapping := range bindableKindRestMappings {
+		// TODO: Debug into why List returns all the versions instead of the GVR version
+		// List all the instances of the restMapping object
+		resources, err := o.kubernetesClient.ListDynamicResources(restMapping.Resource)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, item := range resources.Items {
+			// format: `<name> (<kind>.<group>)`
+			serviceName := fmt.Sprintf("%s (%s.%s)", item.GetName(), item.GetKind(), item.GroupVersionKind().Group)
+			bindableObjectMap[serviceName] = item
+		}
+
+	}
+
+	return bindableObjectMap, nil
+}
