+++ pkg/binding/binding.go
@@ -3,7 +3,6 @@ package binding
 import (
 	"fmt"
 	"path/filepath"
-	"strings"
 
 	bindingApis "github.com/redhat-developer/service-binding-operator/apis"
 	bindingApi "github.com/redhat-developer/service-binding-operator/apis/binding/v1alpha1"
@@ -55,108 +54,6 @@ func (o *BindingClient) GetFlags(flags map[string]string) map[string]string {
 		}
 	}
 	return bindingFlags
-}
-
-// Validate calls Validate method of the adequate backend
-func (o *BindingClient) Validate(flags map[string]string) error {
-	var backend backendpkg.AddBindingBackend
-	if len(flags) == 0 {
-		backend = o.interactiveBackend
-	} else {
-		backend = o.flagsBackend
-	}
-	return backend.Validate(flags)
-}
-
-func (o *BindingClient) SelectServiceInstance(flags map[string]string, serviceMap map[string]unstructured.Unstructured) (string, error) {
-	var backend backendpkg.AddBindingBackend
-	if len(flags) == 0 {
-		backend = o.interactiveBackend
-	} else {
-		backend = o.flagsBackend
-	}
-	return backend.SelectServiceInstance(flags[backendpkg.FLAG_SERVICE], serviceMap)
-}
-
-func (o *BindingClient) AskBindingName(serviceName, componentName string, flags map[string]string) (string, error) {
-	var backend backendpkg.AddBindingBackend
-	if len(flags) == 0 {
-		backend = o.interactiveBackend
-	} else {
-		backend = o.flagsBackend
-	}
-	defaultBindingName := fmt.Sprintf("%v-%v", componentName, serviceName)
-	return backend.AskBindingName(defaultBindingName, flags)
-}
-
-func (o *BindingClient) AskBindAsFiles(flags map[string]string) (bool, error) {
-	var backend backendpkg.AddBindingBackend
-	if len(flags) == 0 {
-		backend = o.interactiveBackend
-	} else {
-		backend = o.flagsBackend
-	}
-	return backend.AskBindAsFiles(flags)
-}
-
-func (o *BindingClient) AddBinding(bindingName string, bindAsFiles bool, unstructuredService unstructured.Unstructured, obj parser.DevfileObj) (parser.DevfileObj, error) {
-	service, err := o.kubernetesClient.NewServiceBindingServiceObject(unstructuredService, bindingName)
-	if err != nil {
-		return obj, err
-	}
-
-	deploymentName := fmt.Sprintf("%s-app", obj.GetMetadataName())
-	deploymentGVR, err := o.kubernetesClient.GetDeploymentAPIVersion()
-	if err != nil {
-		return obj, err
-	}
-
-	serviceBinding := kclient.NewServiceBindingObject(bindingName, bindAsFiles, deploymentName, deploymentGVR, []bindingApi.Mapping{}, []bindingApi.Service{service})
-
-	// Note: we cannot directly marshal the serviceBinding object to yaml because it doesn't do that in the correct k8s manifest format
-	serviceBindingUnstructured, err := kclient.ConvertK8sResourceToUnstructured(serviceBinding)
-	if err != nil {
-		return obj, err
-	}
-	yamlDesc, err := yaml.Marshal(serviceBindingUnstructured.UnstructuredContent())
-	if err != nil {
-		return obj, err
-	}
-
-	return libdevfile.AddKubernetesComponentToDevfile(string(yamlDesc), serviceBinding.Name, obj)
-}
-
-func (o *BindingClient) GetServiceInstances() (map[string]unstructured.Unstructured, error) {
-	// Get the BindableKinds/bindable-kinds object
-	bindableKind, err := o.kubernetesClient.GetBindableKinds()
-	if err != nil {
-		return nil, err
-	}
-
-	// get a list of restMappings of all the GVKs present in bindableKind's Status
-	bindableKindRestMappings, err := o.kubernetesClient.GetBindableKindStatusRestMapping(bindableKind.Status)
-	if err != nil {
-		return nil, err
-	}
-
-	var bindableObjectMap = map[string]unstructured.Unstructured{}
-	for _, restMapping := range bindableKindRestMappings {
-		// TODO: Debug into why List returns all the versions instead of the GVR version
-		// List all the instances of the restMapping object
-		resources, err := o.kubernetesClient.ListDynamicResources(restMapping.Resource)
-		if err != nil {
-			return nil, err
-		}
-
-		for _, item := range resources.Items {
-			// format: `<name> (<kind>.<group>)`
-			serviceName := fmt.Sprintf("%s (%s.%s)", item.GetName(), item.GetKind(), item.GroupVersionKind().Group)
-			bindableObjectMap[serviceName] = item
-		}
-
-	}
-
-	return bindableObjectMap, nil
 }
 
 // GetBindingsFromDevfile returns all ServiceBinding resources declared as Kubernertes component from a Devfile
@@ -331,44 +228,4 @@ func (o *BindingClient) getStatusFromSpec(name string) (*api.ServiceBindingStatu
 		BindingFiles:   bindingFiles,
 		BindingEnvVars: bindingEnvVars,
 	}, nil
-}
-
-// ValidateRemoveBinding validates if the command has adequate arguments/flags
-func (o *BindingClient) ValidateRemoveBinding(flags map[string]string) error {
-	if flags[backendpkg.FLAG_NAME] == "" {
-		return fmt.Errorf("you must specify the service binding name with --%s flag", backendpkg.FLAG_NAME)
-	}
-	return nil
-}
-
-// RemoveBinding removes the binding from devfile
-func (o *BindingClient) RemoveBinding(servicebindingName string, obj parser.DevfileObj) (parser.DevfileObj, error) {
-	var componentName string
-	var options []string
-	components, err := obj.Data.GetComponents(parsercommon.DevfileOptions{
-		ComponentOptions: parsercommon.ComponentOptions{ComponentType: devfilev1alpha2.KubernetesComponentType},
-	})
-	if err != nil {
-		return obj, err
-	}
-	for _, component := range components {
-		var unstructuredObj unstructured.Unstructured
-		unstructuredObj, err = libdevfile.GetK8sComponentAsUnstructured(obj, component.Name, filepath.Dir(obj.Ctx.GetAbsPath()), devfilefs.DefaultFs{})
-		if err != nil {
-			continue
-		}
-		if unstructuredObj.GetKind() == kclient.ServiceBindingKind {
-			options = append(options, unstructuredObj.GetName())
-			if unstructuredObj.GetName() == servicebindingName {
-				componentName = component.Name
-				break
-			}
-		}
-	}
-
-	err = obj.Data.DeleteComponent(componentName)
-	if err != nil {
-		err = fmt.Errorf("Failed to remove Service Binding %q from the the devfile. Available Service Bindings: %s", servicebindingName, strings.Join(options, ", "))
-	}
-	return obj, err
 }
