+++ pkg/binding/remove_binding.go
@@ -1,0 +1,59 @@
+package binding
+
+import (
+	"fmt"
+	"path/filepath"
+	"strings"
+
+	"github.com/devfile/api/v2/pkg/apis/workspaces/v1alpha2"
+	"github.com/devfile/library/pkg/devfile/parser"
+	"github.com/devfile/library/pkg/devfile/parser/data/v2/common"
+	devfilefs "github.com/devfile/library/pkg/testingutil/filesystem"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+
+	backendpkg "github.com/redhat-developer/odo/pkg/binding/backend"
+	"github.com/redhat-developer/odo/pkg/kclient"
+	"github.com/redhat-developer/odo/pkg/libdevfile"
+)
+
+// ValidateRemoveBinding validates if the command has adequate arguments/flags
+func (o *BindingClient) ValidateRemoveBinding(flags map[string]string) error {
+	if flags[backendpkg.FLAG_NAME] == "" {
+		return fmt.Errorf("you must specify the service binding name with --%s flag", backendpkg.FLAG_NAME)
+	}
+	return nil
+}
+
+// RemoveBinding removes the binding from devfile
+func (o *BindingClient) RemoveBinding(servicebindingName string, obj parser.DevfileObj) (parser.DevfileObj, error) {
+	var componentName string
+	var options []string
+	// Get all the K8s type devfile components
+	components, err := obj.Data.GetComponents(common.DevfileOptions{
+		ComponentOptions: common.ComponentOptions{ComponentType: v1alpha2.KubernetesComponentType},
+	})
+	if err != nil {
+		return obj, err
+	}
+	for _, component := range components {
+		var unstructuredObj unstructured.Unstructured
+		// Parse the K8s manifest
+		unstructuredObj, err = libdevfile.GetK8sComponentAsUnstructured(obj, component.Name, filepath.Dir(obj.Ctx.GetAbsPath()), devfilefs.DefaultFs{})
+		if err != nil {
+			continue
+		}
+		if unstructuredObj.GetKind() == kclient.ServiceBindingKind {
+			options = append(options, unstructuredObj.GetName())
+			if unstructuredObj.GetName() == servicebindingName {
+				componentName = component.Name
+				break
+			}
+		}
+	}
+	if componentName == "" {
+		return obj, fmt.Errorf("Service Binding %q not found in the devfile. Available Service Bindings: %s", servicebindingName, strings.Join(options, ", "))
+	}
+
+	err = obj.Data.DeleteComponent(componentName)
+	return obj, err
+}
